Server/requirements.txt
bcrypt==3.1.7
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
psycopg2==2.8.4
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
uuid==1.30
